requirements.txt
python-telegram-bot==20.7
openai==1.3.0
yt-dlp==2024.3.10
spotipy==2.23.0
python-dotenv==1.0.0
aiohttp==3.9.1
pymongo==4.6.0
motor==3.3.2
httpx==0.25.2
pyrogram==2.0.106
tgcrypto==1.2.5
